Fix layout problem with big code blocks on mobile screens
The page layout is kind of messed up for the blog because the main page
uses a flex layout in order to layout child containers but the blog does
not need a flex layout. This results in layout problems when big code
blocks that overflow their parent are rendered on small screens. The
workaround for this is to remove flex classes from the blog slug
template.

There's still an issue with margins of blog pages: There is no
horizontal margin, but there should really be one at least on mobile
screens. This needs to be fixed next.